+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 
  - Add listener and do_not_run_periodically as Function module options
+ - Add Settings Import Log
 
 
 ## 8.15.4 (2023-01-19)